+++ credit-management/src/main/java/org/grnet/creditmanagement/exceptions/RatingPolicyBeforeEarliestException.java
@@ -1,0 +1,7 @@
+package org.grnet.creditmanagement.exceptions;
+
+public class RatingPolicyBeforeEarliestException extends RuntimeException {
+    public RatingPolicyBeforeEarliestException(String message) {
+        super(message);
+    }
+}