+++ CHANGELOG.md
@@ -30,6 +30,8 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the
 
 - [#290](https://github.com/ARGOeu/argo-accounting/pull/290) - ACC-1873 Make
 - credit management module optional via Maven profile.
+- [#291](https://github.com/ARGOeu/argo-accounting/pull/290) - ACC-1866 Add
+- rate-policy endpoint for installation metrics.
 
 ## 1.8.1 - 2025-12-09